pyproject.toml
[project]
name = "edumotion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "colorama>=0.4.6",
    "debugpy>=1.8.13",
    "log-symbols>=0.0.14",
    "numpy>=2.2.4",
    "pyqt6>=6.8.1",
    "pyside6>=6.8.2.1",
    "pytest-benchmark>=5.1.0",
    "skia-python>=87.7",
    "vprof>=0.38",
]

[tool.pyright]
include = ["src"]
exclude = ["**/node_modules",
    "**/__pycache__",
    "src/experimental",
    "src/typestubs"
]
pythonVersion = "3.12"
reportMissingTypeStubs = false
allowUntypedLibraries = ["skia"]
reportExplicitAny = false
reportUnknownMemberType = false
reportAny = false
reportUnknownVariableType = false
